[PV1 Neo] Refactor windows; organize folder strcuture
