+++ Neo/Parcel.Neo/Parcel.Neo.csproj.user
@@ -13,7 +13,7 @@
     <Page Update="PopupWindows\AboutWindow.xaml">
       <SubType>Designer</SubType>
     </Page>
-    <Page Update="ScrollbarResource.xaml">
+    <Page Update="Styling\ScrollbarResource.xaml">
       <SubType>Designer</SubType>
     </Page>
     <Page Update="Styling\NeoMenuStylingDictionary.xaml">
