+++ Neo/Parcel.Neo/PopupWindows/CommentWindow.xaml
@@ -1,9 +1,10 @@
-﻿<local:BaseWindow x:Class="Parcel.Neo.CommentWindow"
+﻿<base:BaseWindow x:Class="Parcel.Neo.PopupWindows.CommentWindow"
                   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
                   xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-                  xmlns:local="clr-namespace:Parcel.Neo"
+                  xmlns:local="clr-namespace:Parcel.Neo.PopupWindows"
+                  xmlns:base="clr-namespace:Parcel.Neo"
                   xmlns:converters="clr-namespace:Parcel.Neo.Converters"
                   xmlns:framework="clr-namespace:Parcel.Neo.Base.Framework;assembly=Parcel.Neo.Base"
                   mc:Ignorable="d"
@@ -23,4 +24,4 @@
                      MinHeight="300"></TextBox>
         </DockPanel>
     </DockPanel>
-</local:BaseWindow>
+</base:BaseWindow>
