+++ Neo/Parcel.Neo/PopupWindows/FunctionsTray.xaml
@@ -1,9 +1,10 @@
-﻿<local:BaseWindow x:Class="Parcel.Neo.PopupTab"
+﻿<base:BaseWindow x:Class="Parcel.Neo.PopupWindows.FunctionsTray"
                   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
                   xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-                  xmlns:local="clr-namespace:Parcel.Neo"
+                  xmlns:local="clr-namespace:Parcel.Neo.PopupWindows"
+                  xmlns:base="clr-namespace:Parcel.Neo"
                   mc:Ignorable="d"
                   WindowStyle="None" ResizeMode="NoResize"
                   Title="Tools" Height="300" SizeToContent="WidthAndHeight"
@@ -110,4 +111,4 @@
             </ListView.ItemsPanel>
         </ListView>
     </DockPanel>
-</local:BaseWindow>
+</base:BaseWindow>
