+++ Neo/Parcel.Neo/PreviewWindows/LiveCodePreviewWindow.xaml
@@ -1,9 +1,10 @@
-﻿<Window x:Class="Parcel.Neo.PreviewWindows.LiveCodePreviewWindow"
+﻿<base:BaseWindow x:Class="Parcel.Neo.PreviewWindows.LiveCodePreviewWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:Parcel.Neo.PreviewWindows"
+        xmlns:base="clr-namespace:Parcel.Neo"
         mc:Ignorable="d"
         Title="Live Code Preview (Experimental)" SizeToContent="WidthAndHeight" WindowStartupLocation="CenterOwner" MinWidth="400" MinHeight="300"
         DataContext="{Binding RelativeSource={RelativeSource self}}">
@@ -42,7 +43,7 @@
             <avalonEdit:TextEditor.Resources>
                 <ResourceDictionary>
                     <ResourceDictionary.MergedDictionaries>
-                        <ResourceDictionary Source="../ScrollbarResource.xaml"/>
+                        <ResourceDictionary Source="../Styling/ScrollbarResource.xaml"/>
                     </ResourceDictionary.MergedDictionaries>
                 </ResourceDictionary>
             </avalonEdit:TextEditor.Resources>
@@ -64,4 +65,4 @@
             </avalonEdit:TextEditor.Foreground>
         </avalonEdit:TextEditor>
     </DockPanel>
-</Window>
+</base:BaseWindow>
