+++ Neo/Parcel.Neo/Styling/ScrollbarResource.xaml
